requirements.txt
openai
gradio
scikit-learn
httpx[socks]
python-dotenv